fix completion window handling when quote is inside block.
Fixes the completion window behavior when there's a quote inside a block (e.g. the code `{ "asdf" }`). In this case, the completion window was automatically opening on the closing quotation mark.

This causes an issue because if you type something like `if(cond) { Console.WriteLine("asdf"); }`, the completion window would open on the closing quotation mark, and then commit the first completion when the closing parenthesis is typed. This resulted in `if(cond) { Console.WriteLine("asdf"as); }` where that extra `as` was committed from the completion window.

As far as I can tell, in Visual Studio pressing quote will never open the completion window, so I added it to the list of skipped characters in the `ShouldOpenCompletionWindowAsync` callback.